removed bloatware code, file structure improved
