+++ src/index.css
@@ -1,13 +1,137 @@
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+.page-wrapper{
+  background-color: black;
+  font-family: Cataclysmo;
+  height: 100vh;
+  overflow: hidden;
 }
 
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+.left-section{
+  z-index: 10;
 }
+.left-section .intro{
+  background-color: #fff;
+  border-radius: 5%;
+}
+.left-section .intro h2{
+  font-size: 38px;
+  font-family: Cataclysmo;
+}
+.left-section .intro p{
+  font-size: 15px;
+  text-align: center;
+  font-family: KomikaAxis;
+  color: black;
+}
+.left-section .intro figure{
+  height: 125px;
+  width: 125px;
+  border-radius: 50%;
+  border: 2px solid white;
+  background-color: white;
+  overflow: hidden;
+  transition: all 0.5s;
+}
+.left-section .intro figure:hover{
+  transform: scale(1.05);
+}
+.left-section .intro figure img{
+  width: 100%;
+}
+.nav-list{
+  font-weight: 400;
+  font-family: Cataclysmo;
+  background-color: black;
+}
+.nav-list .single-nav{
+  cursor: pointer;
+  font-size: 25px;
+  border-top: 2px solid white;
+  border-bottom: 2px solid white;
+  color: white;
+  transition: all 0.6s;
+}
+.nav-list .single-nav:hover{
+  background-color: black !important;
+  color: #fff !important;
+}
+.display{
+  background-color: #31313A;
+  border-left: 10px solid black;
+  color: rgba(255, 255, 255, .75);
+}
+
+@media(max-width:991px){
+  .page-wrapper{
+    overflow: scroll;
+    flex-direction: column;
+    animation: none !important;
+    height: 100%;
+  }
+  .display{
+    border: none;
+  }
+}
+/* @media(max-width:576px){
+  .left-section{
+    display: flex;
+  }
+  .intro{
+    flex-direction: column !important;
+    height: 250px !important;
+  }
+} */
+ /* @media (max-width: 991px){
+  .page-wrapper{
+    overflow: scroll;
+    flex-direction: column;
+    animation: none !important;
+  }
+  .mob-display{
+    display: flex;
+    flex-direction: column;
+  }
+  .left-section{
+    display: none !important;
+  }
+  .display{
+    display: none !important;
+  }
+  .mob-background{
+    background-color: white;
+    padding: 20px 0px;
+    height: 100vh;
+  }
+} */
+/*
+@media (max-width:768px){
+  .ham-bar{
+    display: flex !important;
+  }
+  .ham-menu{
+    display: flex;
+  }
+  .left-section{
+    display: none;
+  }
+  .display{
+    height: 100vh;
+    border: 0;
+  }
+	.animated {
+		-o-transition-property: none !important;
+		-moz-transition-property: none !important;
+		-ms-transition-property: none !important;
+		-webkit-transition-property: none !important;
+		transition-property: none !important;
+		-o-transform: none !important;
+		-moz-transform: none !important;
+		-ms-transform: none !important;
+		-webkit-transform: none !important;
+		transform: none !important;
+		-webkit-animation: none !important;
+		-moz-animation: none !important;
+		-o-animation: none !important;
+		-ms-animation: none !important;
+		animation: none !important;
+	}
+} */
